pyproject.toml
[project]
name = "ocr-annotation-dumper"
version = "0.1.0"
description = "Tool to dump Google Cloud Vision OCR predictions from images and documents"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-cloud-vision>=3.5.0",
    "Pillow>=10.0.0",
    "click>=8.0.0",
    "rich>=13.0.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
ocr-dump = "ocr_annotation_dumper.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
